+++ desktop-widgets/divelogexportdialog.h
@@ -34,6 +34,8 @@ private:
 	void exportHtmlInit(const QString &filename);
 	void export_depths(const char *filename, const bool selected_only);
 	void export_TeX(const char *filename, const bool selected_only);
+	void exportProfile(const QString filename, const bool selected_only);
+	void saveProfile(const struct dive *dive, const QString filename);
 };
 
 #endif // DIVELOGEXPORTDIALOG_H